+++ server/log.js
@@ -4,10 +4,10 @@
  * or just plain printing errors to the log.
  */
 
-const Raven = require('raven');
+const Sentry = require('@sentry/node');
 
 const ENV = process.env.REACT_APP_ENV;
-const SENTRY_DSN = process.env.SERVER_SENTRY_DSN;
+const SENTRY_DSN = process.env.REACT_APP_SENTRY_DSN;
 
 /**
  * Set up error loggin. If a Sentry DSN is defined
@@ -19,12 +19,7 @@ exports.setup = () => {
     // exceptions from starting the server etc. but does not catch the
     // ones thrown from Express.js middleware functions. For those
     // an error handler has to be added to the Express app.
-    Raven.config(SENTRY_DSN, {
-      environment: ENV,
-      autoBreadcrumbs: {
-        http: true,
-      },
-    }).install();
+    Sentry.init({ dsn: SENTRY_DSN, environment: ENV });
   }
 };
 
@@ -34,7 +29,7 @@ exports.setup = () => {
  */
 exports.requestHandler = () => {
   if (SENTRY_DSN) {
-    return Raven.requestHandler();
+    return Sentry.Handlers.requestHandler();
   } else {
     return (req, res, next) => {
       next();
@@ -48,7 +43,7 @@ exports.requestHandler = () => {
  */
 exports.errorHandler = () => {
   if (SENTRY_DSN) {
-    return Raven.errorHandler();
+    return Sentry.Handlers.errorHandler();
   } else {
     return (err, req, res, next) => {
       next(err);
@@ -67,7 +62,15 @@ exports.errorHandler = () => {
  */
 exports.error = (e, code, data) => {
   if (SENTRY_DSN) {
-    Raven.captureException(e, { tags: { code }, extra: data });
+    const extra = { ...data, apiErrorData: responseApiErrorInfo(e) };
+
+    Sentry.withScope(scope => {
+      scope.setTag('code', code);
+      Object.keys(extra).forEach(key => {
+        scope.setExtra(key, extra[key]);
+      });
+      Sentry.captureException(e);
+    });
   } else {
     console.error(e);
     console.error(code);