+++ src/config.js
@@ -62,7 +62,7 @@ const currency = process.env.REACT_APP_SHARETRIBE_MARKETPLACE_CURRENCY;
 const listingMinimumPriceSubUnits = 0;
 
 // Sentry DSN (Data Source Name), a client key for authenticating calls to Sentry
-const sentryDsn = process.env.REACT_APP_PUBLIC_SENTRY_DSN;
+const sentryDsn = process.env.REACT_APP_SENTRY_DSN;
 
 // If webapp is using SSL (i.e. it's behind 'https' protocol)
 const usingSSL = process.env.REACT_APP_SHARETRIBE_USING_SSL === 'true';